refactor: Revert unrelated changes to project.ts and messageParser.ts
