+++ src/messageParser.ts
@@ -13,11 +13,12 @@
 import * as vscode from 'vscode';
 import { JSONPath, visit } from 'jsonc-parser';
 import XRegExp = require('xregexp');
-import { locateL10nYaml, parseYaml, resolveTemplateArbPath } from './project';
+import { locateL10nYaml } from './project';
 import { L10nYaml } from './extension';
 import { Diagnostics } from './diagnose';
 import { Decorator } from './decorate';
 import path = require('path');
+import YAML = require('yaml');
 import fs = require('fs');
 
 export class Parser {
@@ -206,8 +207,13 @@ export class Parser {
 			return path.isAbsolute(messageList.templatePath)
 				? messageList.templatePath
 				: path.join(path.dirname(document.uri.fsPath), messageList.templatePath);
-		} else if (l10nOptions !== undefined && l10nYamlPath) {
-			return resolveTemplateArbPath(l10nYamlPath, l10nOptions);
+		} else if (l10nOptions !== undefined) {
+			const templateRootFromOptions = l10nOptions?.['arb-dir'] ?? 'lib/l10n';
+			const templatePathFromOptions = l10nOptions?.['template-arb-file'] ?? 'app_en.arb';
+
+			return path.isAbsolute(templatePathFromOptions)
+				? templatePathFromOptions
+				: path.join(path.dirname(l10nYamlPath), templateRootFromOptions, templatePathFromOptions);
 		}
 	}
 
@@ -276,6 +282,14 @@ function matchCurlyBrackets(v: StringLiteral, l10nOptions?: L10nYaml): MatchRecu
 		values.push(...newLocal.map(l => subLiteral.convertMatch(l)));
 	}
 	return values;
+}
+
+function parseYaml(uri: string): L10nYaml | undefined {
+	if (!fs.existsSync(uri)) {
+		return;
+	}
+	const yaml = fs.readFileSync(uri, "utf8");
+	return YAML.parse(yaml) as L10nYaml;
 }
 
 export function getUnescapedRegions(expression: string): [number, number][] {
